Include CheckStatusFileExistence
1. Include CheckStatusFileExistence to simulate the polling event
Polling event simulation:
a. Assuming other process create the status file when the test is completed.
b. The polling event (CheckStatusFileExistence) will be executed as sub thread at the background, when the status file is generated, the polling event will notice and trigger the FileFound event.
c. When the FileFound event triggered, the UI (winform) will display the changes. For now, the changes will be displayed in MsgTextBox.