Remove the delegate.
1. We found that we can use event to update the ui without encountered the System.InvalidOperationException below:
System.InvalidOperationException: 'Cross-thread operation not valid: Control ''MsgTextBox' accessed from a thread other than the thread it was created on.'
2. The error "System.InvalidOperationException: 'Cross-thread operation not valid: Control 'MsgTextBox' accessed from a thread other than the thread it was created on.'" is a classic threading issue in WinForms. It means that you're trying to update a UI control (in this case, MsgTextBox) from a thread other than the UI thread, which is not allowed.
3. This error is happening because the MonitorFiles() method in your CheckStatusFileExistence class is running on a separate thread, and when it finds the file, it's directly calling the FileFound event, which in turn tries to update MsgTextBox without switching back to the UI thread.
4. Here's how to fix it:
Use Invoke to Marshal the Call to the UI Thread:
You need to use Invoke or BeginInvoke to execute the code that updates textBox1 on the UI thread.
5. Why This Fix Works:
Invoke ensures that the code inside the Action delegate is executed on the UI thread.
By passing the relevant control to the class, you can execute the invoke call from within the thread.
This prevents the cross-thread access error.